mediapipe_motioneditor_plugin/requirements.txt
mediapipe>=0.10.9,<0.10.21
opencv-python>=4.8.0,<4.10
numpy>=1.24.0,<2
Pillow>=10.0.0
tqdm>=4.65.0

